# app/docs/swagger.yaml
basePath: /
definitions:
...
  CalorieGuide-db_internal_food.Food:
    description: Продукт
    properties:
      author_id:
        type: integer
      calories:
        type: integer
      carbohydrates:
        type: integer
      description:
        type: string
      fats:
        type: integer
      food_name:
        type: string
      id:
        type: integer
      likes:
        type: integer
      picture:
        items:
          type: integer
        type: array
      proteins:
        type: integer
    type: object
  CalorieGuide-db_internal_food.WithLike:
    description: Продукт с лайком
    properties:
      author_id:
        type: integer
      calories:
        type: integer
      carbohydrates:
        type: integer
      description:
        type: string
      fats:
        type: integer
      food_name:
        type: string
      id:
        type: integer
      isLiked:
        type: boolean
      likes:
        type: integer
      picture:
        items:
          type: integer
        type: array
      proteins:
        type: integer
    type: object
...
  internal_food_handlers.AddResponse:
    properties:
      product:
        $ref: '#/definitions/CalorieGuide-db_internal_food.Food'
    type: object
  internal_food_handlers.FindAllRequest:
    properties:
      sort:
        type: string
      two-decade:
        type: integer
      user_id:
        type: integer
    type: object
  internal_food_handlers.FindAllResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/CalorieGuide-db_internal_food.WithLike'
        type: array
    type: object
  internal_food_handlers.LikeResponse:
    properties:
      action:
        type: string
      product_id:
        type: integer
      user_id:
        type: integer
    type: object
  internal_food_handlers.SearchRequest:
    properties:
      user:
        type: integer
      word:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: API для работы с продуктами, пользователями и приёмами пищи
  title: CalorieGuide API
  version: "1.0"
paths:
...
  /product:
    post:
      consumes:
      - application/json
      description: Создает новый продукт
      parameters:
      - description: Данные продукта
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CalorieGuide-db_internal_food.Food'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_food_handlers.AddResponse'
        "400":
          description: Ошибка при создании
          schema:
            type: string
        "401":
          description: Нет доступа
          schema:
            type: string
      summary: Добавить новый продукт
      tags:
      - products
  /products:
    post:
      consumes:
      - application/json
      description: Возвращает список продуктов с возможностью сортировки
      parameters:
      - description: Параметры сортировки
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_food_handlers.FindAllRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_food_handlers.FindAllResponse'
        "400":
          description: Ошибка при получении
          schema:
            type: string
      summary: Получить список продуктов
      tags:
      - products
  /products/{id}:
    delete:
      description: Удаляет продукт по ID
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CalorieGuide-db_internal_food.Food'
        "400":
          description: Ошибка при удалении
          schema:
            type: string
        "401":
          description: Нет доступа
          schema:
            type: string
      summary: Удалить продукт
      tags:
      - products
    get:
      description: Возвращает продукт по его ID
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CalorieGuide-db_internal_food.Food'
        "400":
          description: Продукт не найден
          schema:
            type: string
      summary: Получить продукт по ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Обновляет продукт по ID
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные продукта
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CalorieGuide-db_internal_food.Food'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CalorieGuide-db_internal_food.Food'
        "400":
          description: Ошибка при обновлении
          schema:
            type: string
        "401":
          description: Нет доступа
          schema:
            type: string
      summary: Обновить продукт
      tags:
      - products
  /products/like:
    post:
      consumes:
      - application/json
      description: Ставит или убирает лайк продукту
      parameters:
      - description: Данные лайка
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_food_handlers.LikeResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_food_handlers.LikeResponse'
        "400":
          description: Ошибка при лайке
          schema:
            type: string
        "401":
          description: Нет доступа
          schema:
            type: string
      summary: Лайк продукта
      tags:
      - products
  /products/search:
    post:
      consumes:
      - application/json
      description: Ищет продукты по ключевому слову
      parameters:
      - description: Поисковый запрос
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_food_handlers.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_food_handlers.FindAllResponse'
        "400":
          description: Ошибка при поиске
          schema:
            type: string
      summary: Поиск продуктов
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"